+++ src/anemoi/inference/commands/requests.py
@@ -12,6 +12,7 @@ from anemoi.utils.grib import shortname_to_paramid
 
 from ..checkpoint import Checkpoint
 from . import Command
+from .retrieve import checkpoint_to_requests
 
 
 class RequestCmd(Command):
@@ -30,8 +31,7 @@ class RequestCmd(Command):
         raise NotImplementedError("Use command `retrieve` instead.")
 
         c = Checkpoint(args.path)
-
-        for r in c.mars_requests(dates=[-1], use_grib_paramid=args.use_grib_paramid):
+        for r in checkpoint_to_requests(c, date=-1, use_grib_paramid=args.use_grib_paramid):
             if args.mars:
                 req = ["retrieve,target=data"]
                 for k, v in r.items():