+++ src/anemoi/inference/inputs/gribfile.py
@@ -22,9 +22,7 @@ LOG = logging.getLogger(__name__)
 @input_registry.register("grib")
 @main_argument("path")
 class GribFileInput(GribInput):
-    """
-    Handles grib files
-    """
+    """Handles grib files"""
 
     trace_name = "grib file"